+++ README.md
@@ -53,8 +53,8 @@
 <br>
 
 ## 5단계  - 자동차 경주(리팩터링)
-- [ ] MVC 패턴 기반으로 리팩터링 한다.
-  - [ ] 핵심 비즈니스 로직을 domain 패키지로 이동한다.
-  - [ ] UI 관련 객체를 view 패키지로 이동한다.
-  - [ ] view 패키지 객체는 domain 패키지를 의존할 수 있다.
-  - [ ] domain 패키지 객체는 view 패키지 객체를 의존할 수 없다.
+- [x] MVC 패턴 기반으로 리팩터링 한다.
+  - [x] 핵심 비즈니스 로직을 domain 패키지로 이동한다.
+  - [x] UI 관련 객체를 view 패키지로 이동한다.
+  - [x] view 패키지 객체는 domain 패키지를 의존할 수 있다.
+  - [x] domain 패키지 객체는 view 패키지 객체를 의존할 수 없다.
